Move Command into a separate file
